fix(validator): change int to long

+++ StandUsers.Application/Interfaces/IValidateUserUseCase.cs
@@ -3,5 +3,5 @@
 public interface IValidateUserUseCase
 {
     Task<bool> EmailExistsAsync(string email);
-    Task<bool> IdentificationNumberExistsAsync(int identificationNumber);
+    Task<bool> IdentificationNumberExistsAsync(long identificationNumber);
 }
